% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm.R
\name{affinitySequenceSet}
\alias{affinitySequenceSet}
\title{Calculate total affinity over a set of sequences}
\usage{
affinitySequenceSet(scores, seq.len, pwm.len)
}
\arguments{
\item{scores}{affinity scores for individual sequences}

\item{seq.len}{lengths of sequences}

\item{pwm.len}{lengths of PWMs}
}
\description{
Calculate total affinity over a set of sequences
}
