\name{as.data.frame,MotifEnrichmentReport-method}
\alias{as.data.frame}
\alias{as.data.frame,MotifEnrichmentReport-method}
\title{Convert a MotifEnrichmentReport into a data.frame object}
\arguments{
  \item{x}{the MotifEnrichmentReport object}

  \item{row.names}{unused}

  \item{optional}{unused}

  \item{...}{unused}
}
\description{
  Convert a MotifEnrichmentReport into a data.frame object
}

