% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifEnrichmentReport-methods.R
\name{as.data.frame,MotifEnrichmentReport-method}
\alias{as.data.frame,MotifEnrichmentReport-method}
\alias{as.data.frame}
\title{Convert a MotifEnrichmentReport into a data.frame object}
\usage{
\S4method{as.data.frame}{MotifEnrichmentReport}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{the MotifEnrichmentReport object}

\item{row.names}{unused}

\item{optional}{unused}

\item{...}{unused}
}
\description{
Convert a MotifEnrichmentReport into a data.frame object
}
