% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{divideRows}
\alias{divideRows}
\title{Divide each row of a matrix with a vector}
\usage{
divideRows(m, v)
}
\arguments{
\item{m}{matrix to be divided}

\item{v}{the vector to use for division}
}
\description{
Divide each row of a matrix with a vector
}
