% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm.R
\name{.inputParamMotifs}
\alias{.inputParamMotifs}
\title{Normalizes the motifs input argument for multiple functions}
\usage{
.inputParamMotifs(motifs)
}
\arguments{
\item{motifs}{a list of motifs either as frequency matrices (PFM) or as PWM objects. If PFMs are specified
they are converted to PWMs using uniform background.}
}
\description{
Normalizes the motifs input argument for multiple functions
}
