\name{.inputParamSequences}
\alias{.inputParamSequences}
\title{Normalize the sequences input argument...}
\usage{.inputParamSequences(sequences)
}
\description{Normalize the sequences input argument}
\arguments{\item{sequences}{a set of sequences to be scanned, a list of DNAString or other scannable objects}
}

