% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity.R
\name{.inputPFMfromMatrixOrPWM}
\alias{.inputPFMfromMatrixOrPWM}
\title{Check the frequency matrix input parameter for motifSimilarity}
\usage{
.inputPFMfromMatrixOrPWM(m)
}
\arguments{
\item{m}{either a PWM object or a matrix}
}
\value{
corresponding PFM
}
\description{
Check the frequency matrix input parameter for motifSimilarity
}
