% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{.normalize.bg.seq}
\alias{.normalize.bg.seq}
\title{check consistency of bg.seq input parameter}
\usage{
.normalize.bg.seq(bg.seq)
}
\arguments{
\item{bg.seq}{a set of background sequences, either a list of DNAString object or DNAStringSet object}
}
\description{
check consistency of bg.seq input parameter
}
