\name{.normargPfm}
\alias{.normargPfm}
\title{Input parameter normalization for PWMUnscaled...}
\usage{.normargPfm(x)
}
\description{Input parameter normalization for PWMUnscaled}
\details{This function is from Biostrings package.
A Position Frequency Matrix (PFM) is also represented as an ordinary
matrix. Unlike a PWM, it must be of type integer (it will typically be
the result of consensusMatrix()).}
\arguments{\item{x}{a frequency matrix}
}

