\name{.normargPriorParams}
\alias{.normargPriorParams}
\title{Input parameter normalization function for PWMUnscaled...}
\usage{.normargPriorParams(prior.params)
}
\description{Input parameter normalization function for PWMUnscaled}
\details{This function is from Biostrings package}
\arguments{\item{prior.params}{Typical 'prior.params' vector: c(A=0.25, C=0.25, G=0.25, T=0.25)}
}

