% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{getPromoters}
\alias{getPromoters}
\title{Get the promoter sequences either for a named organism such as "dm3" or a BSgenome object}
\usage{
getPromoters(organismOrGenome)
}
\arguments{
\item{organismOrGenome}{either organism name, e.g. "dm3", or BSgenome object}
}
\value{
a list of: promoters - DNAStringSet of (unique) promoters; organism - name of species; version - genome version
}
\description{
Get the promoter sequences either for a named organism such as "dm3" or a BSgenome object
}
