% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm.R
\name{logNormPvalSequenceSet}
\alias{logNormPvalSequenceSet}
\title{Lognormal P-value for a set of sequences}
\usage{
logNormPvalSequenceSet(scores, seq.len, pwm.len, bg.mean, bg.sd, bg.len)
}
\arguments{
\item{scores}{a matrix of per-sequence affinity scores}

\item{seq.len}{lengths of sequences}

\item{pwm.len}{lengths of pwms}

\item{bg.mean}{mean background at length of bg.len}

\item{bg.sd}{standard deviation of background at length of bg.len}

\item{bg.len}{the length for which mean and sd are calculated}
}
\value{
P-value
}
\description{
Lognormal P-value for a set of sequences
}
