% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{makePriors}
\alias{makePriors}
\title{Make priors from background sequences}
\usage{
makePriors(bg.seq, bg.pseudo.count)
}
\arguments{
\item{bg.seq}{a set of background sequences}

\item{bg.pseudo.count}{the total pseudocount shared between nucleotides}
}
\description{
These priors serve both as background nucleotide frequencies and pseudo-counts
for PWMs.
}
\examples{
# some example sequences
sequences = list(DNAString("AAAGAGAGTGACCGATGAC"), DNAString("ACGATGAGGATGAC"))
# make priors with pseudo-count of 1 shared between them
makePriors(sequences, 1)
}
