% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{makeStartEndPos}
\alias{makeStartEndPos}
\title{Divide total.len into fragments of length len by providing start,end positions}
\usage{
makeStartEndPos(total.len, len)
}
\arguments{
\item{total.len}{total available length to be subdivided}

\item{len}{size of the individual chunk}
}
\value{
a data.frame containing paired up start,end positions
}
\description{
Divide total.len into fragments of length len by providing start,end positions
}
