% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataClasses.R
\docType{class}
\name{MotifEnrichmentResults-class}
\alias{MotifEnrichmentResults-class}
\title{A wrapper class for results of motifEnrichment() that should make it easier to access the results.}
\description{
Note that this is only a wrapper around a list which is the return value in PWMEnrich 1.3 and as such it
provides the same interface as a list (for backward compatibility), with some additional methods.
}
\section{Slots}{
 
 \describe{
   \item{\code{res}:}{a list of old results with elements such as: sequence.bg, sequence.nobg, group.bg, group.nobg}
 }
}

