% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm.R
\name{motifPrAUC}
\alias{motifPrAUC}
\title{Calculate PR-AUC for motifs ranked according to some scoring scheme}
\usage{
motifPrAUC(seq.res)
}
\arguments{
\item{seq.res}{a matrix where each column represents a PWM and each row a result for a different sequence.}
}
\description{
Note that this function asssumes that smaller values are better!
}
