\name{motifRecoveryAUC}
\alias{motifRecoveryAUC}
\title{Calculate Recovery-AUC for motifs ranked according to some scoring scheme}
\usage{
  motifRecoveryAUC(seq.res)
}
\arguments{
  \item{seq.res}{a matrix where each column represents a
  PWM and each row a result for a different sequence.}
}
\description{
  Note that this function asssumes that smaller values are
  better!
}

