\name{operators-MotifEnrichmentResults}
\alias{names,MotifEnrichmentResults-method}
\title{Names of variables}
\usage{\S4method{names}{MotifEnrichmentResults}(x)
\S4method{$}{MotifEnrichmentResults}(x, name)
}
\description{Name of different pieces of information associated with MotifEnrichmentResults}
\value{\code{names,MotifEnrichmentResults-method}: the names of the variables

}
\arguments{\item{x}{the MotifEnrichmentResults object}
\item{name}{the variable name}
}
\alias{$,MotifEnrichmentResults-method}
\alias{$,MotifEnrichmentResults-method}

