\name{operators-PWM}
\alias{names,PWM-method}
\title{Names of variables}
\usage{\S4method{names}{PWM}(x)
\S4method{$}{PWM}(x, name)
\S4method{length}{PWM}(x)
}
\description{Name of different pieces of information associated with PWM}
\value{\code{names,PWM-method}: the names of the variables

}
\arguments{\item{x}{the PWM object}
\item{name}{the variable name}
}
\alias{$,PWM-method}
\alias{$,PWM-method}
\alias{length,PWM-method}
\alias{length}
\details{\code{length,PWM-method}: Returns the motif length, i.e. the number of columns in the PWM.

}

