\name{operators-PWMCutoffBackground}
\alias{names,PWMCutoffBackground-method}
\title{Names of variables}
\usage{\S4method{names}{PWMCutoffBackground}(x)
\S4method{$}{PWMCutoffBackground}(x, name)
}
\description{Name of different pieces of information associated with PWMCutoffBackground}
\value{\code{names,PWMCutoffBackground-method}: the names of the variables

}
\arguments{\item{x}{the PWMCutoffBackground object}
\item{name}{the variable name}
}
\alias{$,PWMCutoffBackground-method}
\alias{$,PWMCutoffBackground-method}

