\name{operators-PWMEmpiricalBackground}
\alias{names,PWMEmpiricalBackground-method}
\title{Names of variables}
\usage{\S4method{names}{PWMEmpiricalBackground}(x)
\S4method{$}{PWMEmpiricalBackground}(x, name)
}
\description{Name of different pieces of information associated with PWMEmpiricalBackground}
\value{\code{names,PWMEmpiricalBackground-method}: the names of the variables

}
\arguments{\item{x}{the PWMEmpiricalBackground object}
\item{name}{the variable name}
}
\alias{$,PWMEmpiricalBackground-method}
\alias{$,PWMEmpiricalBackground-method}

