\name{operators-PWMGEVBackground}
\alias{names,PWMGEVBackground-method}
\title{Names of variables}
\usage{\S4method{names}{PWMGEVBackground}(x)
\S4method{$}{PWMGEVBackground}(x, name)
}
\description{Name of different pieces of information associated with PWMGEVBackground}
\value{\code{names,PWMGEVBackground-method}: the names of the variables

}
\arguments{\item{x}{the PWMGEVBackground object}
\item{name}{the variable name}
}
\alias{$,PWMGEVBackground-method}
\alias{$,PWMGEVBackground-method}

