\name{operators-PWMLognBackground}
\alias{names,PWMLognBackground-method}
\title{Names of variables}
\usage{\S4method{names}{PWMLognBackground}(x)
\S4method{$}{PWMLognBackground}(x, name)
}
\description{Name of different pieces of information associated with PWMLognBackground}
\value{\code{names,PWMLognBackground-method}: the names of the variables

}
\arguments{\item{x}{the PWMLognBackground object}
\item{name}{the variable name}
}
\alias{$,PWMLognBackground-method}
\alias{$,PWMLognBackground-method}

