% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifEnrichmentReport-methods.R
\name{names,MotifEnrichmentReport}
\alias{names,MotifEnrichmentReport}
\alias{names,MotifEnrichmentReport-method}
\alias{$,MotifEnrichmentReport-method}
\alias{[,MotifEnrichmentReport-method}
\title{Names of variables}
\usage{
\S4method{names}{MotifEnrichmentReport}(x)

\S4method{$}{MotifEnrichmentReport}(x, name)

\S4method{[}{MotifEnrichmentReport}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{the MotifEnrichmentReport object}

\item{name}{the variable name}

\item{i}{the row selector}

\item{j}{unused}

\item{...}{unused}

\item{drop}{unused (always FALSE)}
}
\value{
the names of the variables
}
\description{
Columns stored in the motif enrichment report
}
