\name{names,MotifEnrichmentResults}
\alias{$,MotifEnrichmentResults-method}
\alias{names,MotifEnrichmentResults}
\alias{names,MotifEnrichmentResults-method}
\title{Names of variables}
\arguments{
  \item{x}{the MotifEnrichmentResults object}

  \item{x}{the MotifEnrichmentResults object}

  \item{name}{the variable name}
}
\value{
  the names of the variables
}
\description{
  Name of different pieces of information associated with
  MotifEnrichmentResults

  Access a property by name
}

