% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWM-methods.R
\name{names,PWM}
\alias{names,PWM}
\alias{names,PWM-method}
\alias{$,PWM-method}
\alias{length,PWM-method}
\title{Names of variables}
\usage{
\S4method{names}{PWM}(x)

\S4method{$}{PWM}(x, name)

\S4method{length}{PWM}(x)
}
\arguments{
\item{x}{the PWM object}

\item{name}{the variable name}
}
\value{
the names of the variables
}
\description{
Name of different pieces of information associated with PWM

Returns the motif length, i.e. the number of columns in the PWM.
}
