\name{names,PWMCutoffBackground}
\alias{names,PWMCutoffBackground}
\alias{names,PWMCutoffBackground-method}
\alias{$,PWMCutoffBackground-method}
\title{Names of variables}
\arguments{
  \item{x}{the PWMCutoffBackground object}

  \item{x}{the PWMCutoffBackground object}

  \item{name}{the variable name}
}
\value{
  the names of the variables
}
\description{
  Name of different pieces of information associated with
  PWMCutoffBackground

  Access a property by name
}

