\name{names,PWMEmpiricalBackground}
\alias{names,PWMEmpiricalBackground}
\alias{names,PWMEmpiricalBackground-method}
\alias{$,PWMEmpiricalBackground-method}
\title{Names of variables}
\arguments{
  \item{x}{the PWMEmpiricalBackground object}

  \item{x}{the PWMEmpiricalBackground object}

  \item{name}{the variable name}
}
\value{
  the names of the variables
}
\description{
  Name of different pieces of information associated with
  PWMEmpiricalBackground

  Access a property by name
}

