% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMBackground-methods.R
\name{names,PWMGEVBackground}
\alias{names,PWMGEVBackground}
\alias{names,PWMGEVBackground-method}
\alias{$,PWMGEVBackground-method}
\title{Names of variables}
\usage{
\S4method{names}{PWMGEVBackground}(x)

\S4method{$}{PWMGEVBackground}(x, name)
}
\arguments{
\item{x}{the PWMGEVBackground object}

\item{name}{the variable name}
}
\value{
the names of the variables
}
\description{
Name of different pieces of information associated with PWMGEVBackground
}
