% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMBackground-methods.R
\name{names,PWMLognBackground}
\alias{names,PWMLognBackground}
\alias{names,PWMLognBackground-method}
\alias{$,PWMLognBackground-method}
\title{Names of variables}
\usage{
\S4method{names}{PWMLognBackground}(x)

\S4method{$}{PWMLognBackground}(x, name)
}
\arguments{
\item{x}{the PWMLognBackground object}

\item{name}{the variable name}
}
\value{
the names of the variables
}
\description{
Name of different pieces of information associated with PWMLognBackground
}
