\name{pickGenome}
\alias{pickGenome}
\title{A helper function to pick a genome for an organism}
\usage{
  pickGenome(organism)
}
\arguments{
  \item{organism}{either organism name (such as "dm3") or a
  BSgenome object}
}
\value{
  a BSgenome object
}
\description{
  A helper function to pick a genome for an organism
}

