% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot,PWM,missing-method}
\alias{plot,PWM,missing-method}
\title{Plotting for the PWM class}
\usage{
\S4method{plot}{PWM,missing}(x, y, ...)
}
\arguments{
\item{x}{the PWM object}

\item{y}{unused}

\item{...}{other parameters to pass to seqLogo's \code{plot} function}
}
\description{
This function produces a sequence logo (via package seqLogo).
}
\examples{
if(requireNamespace("PWMEnrich.Dmelanogaster.background")){
  data(MotifDb.Dmel, package = "PWMEnrich.Dmelanogaster.background")

  # plot the tinman motif from MotifDb
  plot(MotifDb.Dmel[["tin"]])
}
}
