% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotPFM}
\alias{plotPFM}
\title{Plot a PFM (not PWM) using seqLogo}
\usage{
plotPFM(pfm, ...)
}
\arguments{
\item{pfm}{a matrix where rows are the four nucleotides}

\item{...}{additional parameters for plot()}
}
\description{
Plot a PFM (not PWM) using seqLogo
}
