% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataClasses.R
\docType{class}
\name{PWMCutoffBackground-class}
\alias{PWMCutoffBackground-class}
\title{Hit count background distribution for a set of PWMs}
\description{
Hit count background distribution for a set of PWMs
}
\section{Slots}{
 
 \describe{
   \item{\code{bg.source}:}{textual description of where the background distribution is derived from}
   \item{\code{bg.cutoff}:}{the cutoff score used to find significant motif hits (in log2 odds), either a single value or a vector of values}
   \item{\code{bg.P}:}{the density of significant motif hits per nucleotide in background}
   \item{\code{pwms}:}{the pwms for which the background has been compiled}
 }
}

