% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMEnrich-package.R
\docType{package}
\name{PWMEnrich-package}
\alias{PWMEnrich}
\alias{PWMEnrich-package}
\title{PWMEnrich: PWM enrichment analysis}
\description{
A toolkit of high-level functions for DNA motif scanning
        and enrichment analysis built upon Biostrings. The main
        functionality is PWM enrichment analysis of already known PWMs
        (e.g. from databases such as MotifDb), but the package also
        implements high-level functions for PWM scanning and
        visualisation. The package does not perform "de novo" motif
        discovery, but is instead focused on using motifs that are
        either experimentally derived or computationally constructed by
        other tools.
}
\keyword{internal}
