% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataClasses.R
\docType{class}
\name{PWMLognBackground-class}
\alias{PWMLognBackground-class}
\title{Lognormal background distribution for a set of PWMs}
\description{
Lognormal background distribution for a set of PWMs
}
\section{Slots}{
 
 \describe{
   \item{\code{bg.source}:}{textual description of where the background distribution is derived from}
   \item{\code{bg.len}:}{the length to which the background is normalized to. This is a vector of values, can have a different value for each motif.}
   \item{\code{bg.mean}:}{the mean value of the lognormal distribution at bg.len}
   \item{\code{bg.sd}:}{the standard deviation of the lognormal distribution at bg.len}
   \item{\code{pwms}:}{the pwms for which the background has been compiled}
 }
}

