\name{readTRANSFAC}
\alias{readTRANSFAC}
\title{Read in motifs in TRANSFAC format...}
\usage{readTRANSFAC(file, remove.acc=TRUE)
}
\description{Read in motifs in TRANSFAC format}
\value{a list of matrices representing motifs (with four nucleotides as rows)}
\arguments{\item{file}{the filename}
\item{remove.acc}{if to ignore transfac accession numbers}
}

