% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWM-methods.R
\name{reverseComplement,PWM-method}
\alias{reverseComplement,PWM-method}
\title{Reverse complement for the PWM object}
\usage{
\S4method{reverseComplement}{PWM}(x, ...)
}
\arguments{
\item{x}{an object of type PWM}

\item{...}{unused}
}
\value{
an object of type PWM that is reverse complement of x
}
\description{
Finds the reverse complement of the PWM
}
\examples{

if(requireNamespace("PWMEnrich.Dmelanogaster.background")){
   data(MotifDb.Dmel.PFM)

   reverseComplement(MotifDb.Dmel.PFM[["ttk"]]) # reverse complement of the ttk PWM
}

}
