\name{selectPromoters}
\alias{selectPromoters}
\title{A helper function to select a set of non-redundant promoters from a genome object}
\usage{
  selectPromoters(genome)
}
\arguments{
  \item{genome}{a BSgenome object}
}
\value{
  a vector of inidicies of non-redundant promoters
}
\description{
  A helper function to select a set of non-redundant
  promoters from a genome object
}

