\name{show,MotifEnrichmentResults-method}
\alias{show,MotifEnrichmentResults-method}
\title{show method for MotifEnrichmentResults...}
\usage{\S4method{show}{MotifEnrichmentResults}(object)
}
\description{show method for MotifEnrichmentResults}
\arguments{\item{object}{the MotifEnrichmentResults object}
}

