\name{show,PWMEmpiricalBackground-method}
\alias{show,PWMEmpiricalBackground-method}
\title{show method for PWMEmpiricalBackground...}
\usage{\S4method{show}{PWMEmpiricalBackground}(object)
}
\description{show method for PWMEmpiricalBackground}
\arguments{\item{object}{the PWMEmpiricalBackground object}
}

