% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifEnrichmentReport-methods.R
\name{show,MotifEnrichmentReport-method}
\alias{show,MotifEnrichmentReport-method}
\title{show method for MotifEnrichmentReport}
\usage{
\S4method{show}{MotifEnrichmentReport}(object)
}
\arguments{
\item{object}{the MotifEnrichmentReport object}
}
\description{
show method for MotifEnrichmentReport
}
