% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifEnrichmentResults-methods.R
\name{show,MotifEnrichmentResults-method}
\alias{show,MotifEnrichmentResults-method}
\title{show method for MotifEnrichmentResults}
\usage{
\S4method{show}{MotifEnrichmentResults}(object)
}
\arguments{
\item{object}{the MotifEnrichmentResults object}
}
\description{
show method for MotifEnrichmentResults
}
