% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMBackground-methods.R
\name{show,PWMCutoffBackground-method}
\alias{show,PWMCutoffBackground-method}
\title{show method for PWMCutoffBackground}
\usage{
\S4method{show}{PWMCutoffBackground}(object)
}
\arguments{
\item{object}{the PWMCutoffBackground object}
}
\description{
show method for PWMCutoffBackground
}
