% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMBackground-methods.R
\name{show,PWMEmpiricalBackground-method}
\alias{show,PWMEmpiricalBackground-method}
\title{show method for PWMEmpiricalBackground}
\usage{
\S4method{show}{PWMEmpiricalBackground}(object)
}
\arguments{
\item{object}{the PWMEmpiricalBackground object}
}
\description{
show method for PWMEmpiricalBackground
}
