% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMBackground-methods.R
\name{show,PWMGEVBackground-method}
\alias{show,PWMGEVBackground-method}
\title{show method for PWMGEVBackground}
\usage{
\S4method{show}{PWMGEVBackground}(object)
}
\arguments{
\item{object}{the PWMGEVBackground object}
}
\description{
show method for PWMGEVBackground
}
