% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMBackground-methods.R
\name{show,PWMLognBackground-method}
\alias{show,PWMLognBackground-method}
\title{show method for PWMLognBackground}
\usage{
\S4method{show}{PWMLognBackground}(object)
}
\arguments{
\item{object}{the PWMLognBackground object}
}
\description{
show method for PWMLognBackground
}
