\name{show}
\title{show method for MotifEnrichmentResults}
\arguments{
  \item{object}{the MotifEnrichmentResults object}

  \item{object}{the MotifEnrichmentReport object}

  \item{object}{the PWMLognBackground object}

  \item{object}{the PWMCutoffBackground object}

  \item{object}{the PWMEmpiricalBackground object}

  \item{object}{the PWMGEVBackground object}

  \item{object}{the PWM object}
}
\description{
  show method for MotifEnrichmentResults

  show method for MotifEnrichmentReport

  show method for PWMLognBackground

  show method for PWMCutoffBackground

  show method for PWMEmpiricalBackground

  show method for PWMGEVBackground

  show method for PWM
}

