% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMBackground-methods.R
\name{[,PWMCutoffBackground-method}
\alias{[,PWMCutoffBackground-method}
\title{Get the background for a subset of PWMs}
\usage{
\S4method{[}{PWMCutoffBackground}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{the PWMCutoffBackground object}

\item{i}{the indicies of PWMs}

\item{j}{unused}

\item{...}{unused}

\item{drop}{unused}
}
\description{
Get the background for a subset of PWMs
}
