\name{subsetting-PWMGEVBackground}
\alias{[,PWMGEVBackground-method}
\title{Get the background for a subset of PWMs}
\arguments{
  \item{x}{the PWMGEVBackground object}

  \item{i}{the indicies of PWMs}

  \item{j}{unused}

  \item{...}{unused}

  \item{drop}{unused}
}
\description{
  Get the background for a subset of PWMs
}

