% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMBackground-methods.R
\name{[,PWMLognBackground-method}
\alias{[,PWMLognBackground-method}
\title{Get the background for a subset of PWMs}
\usage{
\S4method{[}{PWMLognBackground}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{the PWMLognBackground object}

\item{i}{the indicies of PWMs}

\item{j}{unused}

\item{...}{unused}

\item{drop}{unused}
}
\description{
Get the background for a subset of PWMs
}
