\name{affinitySequenceSet}
\alias{affinitySequenceSet}
\title{Calculate total affinity over a set of sequences...}
\usage{affinitySequenceSet(scores, seq.len, pwm.len)
}
\description{Calculate total affinity over a set of sequences}
\arguments{\item{scores}{affinity scores for individual sequences}
\item{seq.len}{lengths of sequences}
\item{pwm.len}{lengths of PWMs}
}

