% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{concatenateSequences}
\alias{concatenateSequences}
\title{Concatenata DNA sequences into a single character object}
\usage{
concatenateSequences(sequences)
}
\arguments{
\item{sequences}{either a list of DNAString objects, or a DNAStringSet}
}
\value{
a single character string
}
\description{
Concatenata DNA sequences into a single character object
}
