\name{concatenateSequences}
\alias{concatenateSequences}
\title{Concatenata DNA sequences into a single character object...}
\usage{concatenateSequences(sequences)
}
\description{Concatenata DNA sequences into a single character object}
\value{a single character string}
\arguments{\item{sequences}{either a list of DNAString objects, or a DNAStringSet}
}

