\name{DNAStringSetToList}
\alias{DNAStringSetToList}
\title{Convert DNAStringSet to list of DNAString objects...}
\usage{DNAStringSetToList(x)
}
\description{Convert DNAStringSet to list of DNAString objects}
\details{as.list doesn't seem to always work for DNAStringSets, so
implementing this ourselves.}
\arguments{\item{x}{an object of class DNAStringSet}
}

