\name{.inputParamMotifs}
\alias{.inputParamMotifs}
\title{Normalizes the motifs input argument for multiple functions...}
\usage{.inputParamMotifs(motifs)
}
\description{Normalizes the motifs input argument for multiple functions}
\arguments{\item{motifs}{a list of motifs either as frequency matrices (PFM) or as PWM objects. If PFMs are specified
they are converted to PWMs using uniform background.}
}

