% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm.R
\name{.inputParamSequences}
\alias{.inputParamSequences}
\title{Normalize the sequences input argument}
\usage{
.inputParamSequences(sequences)
}
\arguments{
\item{sequences}{a set of sequences to be scanned, a list of DNAString or other scannable objects}
}
\description{
Normalize the sequences input argument
}
