\name{.inputPFMfromMatrixOrPWM}
\alias{.inputPFMfromMatrixOrPWM}
\title{Check the frequency matrix input parameter for motifSimilarity...}
\usage{.inputPFMfromMatrixOrPWM(m)
}
\description{Check the frequency matrix input parameter for motifSimilarity}
\value{corresponding PFM}
\arguments{\item{m}{either a PWM object or a matrix}
}

