% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm.R
\name{.normargPfm}
\alias{.normargPfm}
\title{Input parameter normalization for PWMUnscaled}
\usage{
.normargPfm(x)
}
\arguments{
\item{x}{a frequency matrix}
}
\description{
This function is from Biostrings package.
A Position Frequency Matrix (PFM) is also represented as an ordinary
matrix. Unlike a PWM, it must be of type integer (it will typically be
the result of consensusMatrix()).
}
