% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm.R
\name{.normargPriorParams}
\alias{.normargPriorParams}
\title{Input parameter normalization function for PWMUnscaled}
\usage{
.normargPriorParams(prior.params)
}
\arguments{
\item{prior.params}{Typical 'prior.params' vector: c(A=0.25, C=0.25, G=0.25, T=0.25)}
}
\description{
This function is from Biostrings package
}
