\name{makeStartEndPos}
\alias{makeStartEndPos}
\title{Divide total...}
\usage{makeStartEndPos(total.len, len)
}
\description{Divide total.len into fragments of length len by providing start,end positions}
\value{a data.frame containing paired up start,end positions}
\arguments{\item{total.len}{total available length to be subdivided}
\item{len}{size of the individual chunk}
}

