\name{motifPrAUC}
\alias{motifPrAUC}
\title{Calculate PR-AUC for motifs ranked according to some scoring scheme...}
\usage{motifPrAUC(seq.res)
}
\description{Calculate PR-AUC for motifs ranked according to some scoring scheme}
\details{Note that this function asssumes that smaller values are better!}
\arguments{\item{seq.res}{a matrix where each column represents a PWM and each row a result for a different sequence.}
}

