\name{motifRecoveryAUC}
\alias{motifRecoveryAUC}
\title{Calculate Recovery-AUC for motifs ranked according to some scoring scheme...}
\usage{motifRecoveryAUC(seq.res)
}
\description{Calculate Recovery-AUC for motifs ranked according to some scoring scheme}
\details{Note that this function asssumes that smaller values are better!}
\arguments{\item{seq.res}{a matrix where each column represents a PWM and each row a result for a different sequence.}
}

