\name{names,MotifEnrichmentReport}
\alias{[,MotifEnrichmentReport-method}
\alias{$,MotifEnrichmentReport-method}
\alias{names,MotifEnrichmentReport}
\alias{names,MotifEnrichmentReport-method}
\title{Names of variables}
\arguments{
  \item{x}{the MotifEnrichmentReport object}

  \item{x}{the MotifEnrichmentReport object}

  \item{name}{the variable name}

  \item{x}{the MotifEnrichmentReport object}

  \item{i}{the row selector}

  \item{j}{unused}

  \item{...}{unused}

  \item{drop}{unused (always FALSE)}
}
\value{
  the names of the variables
}
\description{
  Columns stored in the motif enrichment report

  Access a column by name

  Subset the report
}

