% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifEnrichmentResults-methods.R
\name{names,MotifEnrichmentResults}
\alias{names,MotifEnrichmentResults}
\alias{names,MotifEnrichmentResults-method}
\alias{$,MotifEnrichmentResults-method}
\title{Names of variables}
\usage{
\S4method{names}{MotifEnrichmentResults}(x)

\S4method{$}{MotifEnrichmentResults}(x, name)
}
\arguments{
\item{x}{the MotifEnrichmentResults object}

\item{name}{the variable name}
}
\value{
the names of the variables
}
\description{
Name of different pieces of information associated with MotifEnrichmentResults
}
