\name{names,PWM}
\alias{length}
\alias{length,PWM-method}
\alias{names,PWM}
\alias{names,PWM-method}
\alias{$,PWM-method}
\title{Names of variables}
\arguments{
  \item{x}{the PWM object}

  \item{x}{the PWM object}

  \item{name}{the variable name}

  \item{x}{the PWM object}
}
\value{
  the names of the variables
}
\description{
  Name of different pieces of information associated with
  PWM

  Access a property by name

  Returns the motif length, i.e. the number of columns in
  the PWM.
}

