% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMBackground-methods.R
\name{names,PWMCutoffBackground}
\alias{names,PWMCutoffBackground}
\alias{names,PWMCutoffBackground-method}
\alias{$,PWMCutoffBackground-method}
\title{Names of variables}
\usage{
\S4method{names}{PWMCutoffBackground}(x)

\S4method{$}{PWMCutoffBackground}(x, name)
}
\arguments{
\item{x}{the PWMCutoffBackground object}

\item{name}{the variable name}
}
\value{
the names of the variables
}
\description{
Name of different pieces of information associated with PWMCutoffBackground
}
