% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMBackground-methods.R
\name{names,PWMEmpiricalBackground}
\alias{names,PWMEmpiricalBackground}
\alias{names,PWMEmpiricalBackground-method}
\alias{$,PWMEmpiricalBackground-method}
\title{Names of variables}
\usage{
\S4method{names}{PWMEmpiricalBackground}(x)

\S4method{$}{PWMEmpiricalBackground}(x, name)
}
\arguments{
\item{x}{the PWMEmpiricalBackground object}

\item{name}{the variable name}
}
\value{
the names of the variables
}
\description{
Name of different pieces of information associated with PWMEmpiricalBackground
}
