\name{names,PWMGEVBackground}
\alias{names,PWMGEVBackground}
\alias{names,PWMGEVBackground-method}
\alias{$,PWMGEVBackground-method}
\title{Names of variables}
\arguments{
  \item{x}{the PWMGEVBackground object}

  \item{x}{the PWMGEVBackground object}

  \item{name}{the variable name}
}
\value{
  the names of the variables
}
\description{
  Name of different pieces of information associated with
  PWMGEVBackground

  Access a property by name
}

