\name{names,PWMLognBackground}
\alias{names,PWMLognBackground}
\alias{names,PWMLognBackground-method}
\alias{$,PWMLognBackground-method}
\title{Names of variables}
\arguments{
  \item{x}{the PWMLognBackground object}

  \item{x}{the PWMLognBackground object}

  \item{name}{the variable name}
}
\value{
  the names of the variables
}
\description{
  Name of different pieces of information associated with
  PWMLognBackground

  Access a property by name
}

