\name{[,PWMCutoffBackground-method}
\alias{[,PWMCutoffBackground-method}
\title{Get the background for a subset of PWMs}
\arguments{
  \item{x}{the PWMCutoffBackground object}

  \item{i}{the indicies of PWMs}

  \item{j}{unused}

  \item{...}{unused}

  \item{drop}{unused}
}
\description{
  Get the background for a subset of PWMs
}

