\name{subsetting-PWMEmpiricalBackground}
\alias{[,PWMEmpiricalBackground-method}
\title{Get the background for a subset of PWMs}
\arguments{
  \item{x}{the PWMEmpiricalBackground object}

  \item{i}{the indicies of PWMs}

  \item{j}{unused}

  \item{...}{unused}

  \item{drop}{unused}
}
\description{
  Get the background for a subset of PWMs
}

